refactor: extract Bedrock gateway transport into its own module
