+++ src/agents/helpers/claudeHandler.ts
@@ -1,6 +1,3 @@
-import logger from '../../config/logger.js'
-import config from '../../config/config.js'
-
 // Helper to determine if Bedrock Claude format should be used
 export function shouldUseClaudeFormat(model: string | undefined, platform: string | undefined): boolean {
   if (platform && typeof platform === 'string') {
@@ -128,83 +125,4 @@ export function transformPayloadForClaude(bodyContent: unknown, defaultLLMModel:
     temperature,
     tools
   })
-}
-
-// Reject ids with characters that could alter the URL path (e.g. "/", "?", "#"); the colon is allowed.
-const BEDROCK_MODEL_ID_PATTERN = /^[A-Za-z0-9._:-]+$/
-
-/**
- * Builds the Bedrock V2 invoke URL for a model: {baseUrl}/model/{modelId}/invoke.
- *
- * The model id goes straight into the URL path and can come from user-set config, so this
- * validates it against the known Bedrock character set first and rejects anything else. It
- * keeps the colon in dated ids (for example ...-v1:0) raw to match HUIT's documented URLs.
- *
- * @throws if the model id contains characters that are not valid in a Bedrock model id.
- */
-export function buildBedrockInvokeUrl(baseUrl: string, modelId: string): string {
-  if (!BEDROCK_MODEL_ID_PATTERN.test(modelId)) {
-    throw new Error(`Invalid Bedrock model id: "${modelId}"`)
-  }
-  // Drop any trailing slash so a base url like ".../v2/" does not produce a double slash.
-  const normalizedBaseUrl = baseUrl.replace(/\/+$/, '')
-  return `${normalizedBaseUrl}/model/${modelId}/invoke`
-}
-
-// Create a custom fetch function for Bedrock Claude or legacy LLM
-export function createClaudeFetchFn(defaultLLMModel: string, defaultLLMPlatform: string) {
-  return async function fetchFn(url: string, init: Parameters<typeof fetch>[1]) {
-    const fetchImpl = async () => {
-      try {
-        let bodyContent: unknown = {}
-        if (init?.body) {
-          if (
-            typeof init.body === 'string' &&
-            (init.body.trim().toLowerCase().startsWith('<!doctype') || init.body.trim().toLowerCase().startsWith('<html'))
-          ) {
-            logger.error('init.body appears to be HTML, not JSON:', init.body)
-            throw new Error('init.body is HTML, not JSON')
-          }
-          try {
-            bodyContent = JSON.parse(init.body as string)
-          } catch (err) {
-            logger.error('init.body is not valid JSON:', init.body)
-            throw err
-          }
-        }
-
-        // Transform payload for Claude if needed
-        bodyContent = transformPayloadForClaude(bodyContent, defaultLLMModel, defaultLLMPlatform)
-
-        // V2 sends the native Bedrock body directly, with no V1 { body, modelId, ... } envelope.
-        const bodyString = JSON.stringify(bodyContent)
-        const modifiedInit = {
-          ...(init || {}),
-          body: bodyString,
-          headers: {
-            'Content-Type': 'application/json',
-            'x-api-key': config.llms.bedrock.key
-          }
-        }
-        // V2 routes by model id in the path; buildBedrockInvokeUrl validates the id first.
-        const invokeUrl = buildBedrockInvokeUrl(config.llms.bedrock.baseUrl, defaultLLMModel)
-        const response = await fetch(invokeUrl, modifiedInit)
-        const contentType = response.headers.get('content-type') || ''
-        if (!response.ok) {
-          const responseText = await response.text()
-          logger.error('Error response Text from proxy:', responseText)
-          throw new Error(`Bedrock proxy error: ${response.status} ${response.statusText}\n${responseText}`)
-        }
-        if (contentType.includes('application/json')) {
-          return response
-        }
-        const responseText = await response.text()
-        throw new Error(`Expected JSON from Bedrock proxy, got ${contentType}: ${responseText}`)
-      } catch (err) {
-        logger.error('Error in fetchFn:', err)
-        throw err
-      }
-    }
-    return fetchImpl()
-  }
 }
